+++ test/index/linking_toc_test.js
@@ -27,7 +27,11 @@ describe('index/linking toc:', function () {
   })
 
   it('renames index.md', function () {
-    expect(this.files['index.html']).toExist()
     expect(this.files['index-2.html']).toExist()
+  })
+
+  it('creates slugs', function () {
+    expect(this.files['index.html'].slug).toEqual('index')
+    expect(this.files['index-2.html'].slug).toEqual('index-2')
   })
 })